moving classes to their own file
