+++ spicelib/__init__.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
 
 # Convenience direct imports
-from .raw.raw_read import RawRead, PlotData, SpiceReadException
+from .raw.raw_read import RawRead, SpiceReadException
+from .raw.plot_data import PlotData
 from .raw.raw_write import RawWrite, Trace
 from .editor.spice_editor import SpiceEditor, SpiceCircuit, SpiceComponent
 from .editor.asc_editor import AscEditor
